fix: use float64 for ModelLimit fields to match OpenAPI spec
The spec defines limit.context and limit.output as "type": "number",
which maps to float64 in Go. int64 would silently truncate fractional
values from the server.

+++ app.go
@@ -67,8 +67,8 @@ type ModelCost struct {
 }
 
 type ModelLimit struct {
-	Context int64 `json:"context"`
-	Output  int64 `json:"output"`
+	Context float64 `json:"context"`
+	Output  float64 `json:"output"`
 }
 
 type ModelModalities struct {
